Add files via upload
This is a Deep Neural Network based IDS or cyber-attack-detection system. It's based on the CICIDS0217 dataset. the system is highly efficient with an accuracy and f1-score of 99.8 %. It is a versatile system which can be used with real-world traffic: it trains on the given data, sets up the necessary algorithms to apply them to new data and visualizes the network situation. It also has option for checking type I & II errors which basically have a low appearance rate ~ 0.2 %. The system also can be used in a real time process and it's capable of reports composing.